Pipeline: filter as a plugin/action
