+++ pydatview/GUITools.py
@@ -187,274 +187,6 @@ class OutlierToolPanel(GUIToolPanel):
 
 
 # --------------------------------------------------------------------------------}
-# --- Filter Tool 
-# --------------------------------------------------------------------------------{
-class FilterToolPanel(GUIToolPanel):
-    """
-    Moving average/Filters
-    A quick and dirty solution to manipulate plotData
-    I need to think of a better way to do that
-    """
-    def __init__(self, parent):
-        from pydatview.tools.signal_analysis import FILTERS
-        super(FilterToolPanel,self).__init__(parent)
-        self.parent = parent # parent is GUIPlotPanel
-
-        self._DEFAULT_FILTERS=FILTERS 
-
-        # Setting default states to parent
-        if 'Filter' not in self.parent.plotDataOptions.keys():
-            self.parent.plotDataOptions['Filter']=None
-        self._filterApplied = type(self.parent.plotDataOptions['Filter'])==dict
-
-
-        self.btClose = self.getBtBitmap(self,'Close','close',self.destroy)
-        self.btAdd   = self.getBtBitmap(self, 'Add','add'  , self.onAdd)
-        self.btHelp  = self.getBtBitmap(self, 'Help','help', self.onHelp)
-        self.btClear = self.getBtBitmap(self, 'Clear Plot','sun'   , self.onClear)
-        self.btApply = self.getToggleBtBitmap(self,'Apply','cloud',self.onToggleApply)
-        self.btPlot  = self.getBtBitmap(self, 'Plot' ,'chart'  , self.onPlot)
-
-        self.cbTabs     = wx.ComboBox(self, -1, choices=[], style=wx.CB_READONLY)
-
-        self.cbFilters = wx.ComboBox(self, choices=[filt['name'] for filt in self._DEFAULT_FILTERS], style=wx.CB_READONLY)
-        self.lbParamName = wx.StaticText(self, -1, '            :')
-        self.cbFilters.SetSelection(0)
-#         self.tParam = wx.TextCtrl(self, wx.ID_ANY, '', style= wx.TE_PROCESS_ENTER, size=wx.Size(60,-1))
-        self.tParam = wx.SpinCtrlDouble(self, value='11', size=wx.Size(80,-1))
-        self.lbInfo = wx.StaticText( self, -1, '')
-
-
-        # --- Layout
-        btSizer  = wx.FlexGridSizer(rows=3, cols=2, hgap=2, vgap=0)
-        btSizer.Add(self.btClose                , 0, flag = wx.ALL|wx.EXPAND, border = 1)
-        btSizer.Add(self.btClear                , 0, flag = wx.ALL|wx.EXPAND, border = 1)
-        btSizer.Add(self.btAdd                  , 0, flag = wx.ALL|wx.EXPAND, border = 1)
-        btSizer.Add(self.btPlot                 , 0, flag = wx.ALL|wx.EXPAND, border = 1)
-        btSizer.Add(self.btHelp                 , 0, flag = wx.ALL|wx.EXPAND, border = 1)
-        btSizer.Add(self.btApply                , 0, flag = wx.ALL|wx.EXPAND, border = 1)
-
-
-        horzSizerT = wx.BoxSizer(wx.HORIZONTAL)
-        horzSizerT.Add(wx.StaticText(self, -1, 'Table:')  , 0, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.LEFT, 5)
-        horzSizerT.Add(self.cbTabs                        , 0, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.LEFT, 1)
-
-        horzSizer = wx.BoxSizer(wx.HORIZONTAL)
-        horzSizer.Add(wx.StaticText(self, -1, 'Filter:')  , 0, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.LEFT, 5)
-        horzSizer.Add(self.cbFilters                      , 0, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.LEFT, 1)
-        horzSizer.Add(self.lbParamName                    , 0, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.LEFT, 5)
-        horzSizer.Add(self.tParam                         , 0, wx.ALIGN_LEFT|wx.ALIGN_CENTER_VERTICAL|wx.LEFT, 1)
-
-
-        vertSizer = wx.BoxSizer(wx.VERTICAL)
-        vertSizer.Add(self.lbInfo  ,0, flag = wx.LEFT          ,border = 5)
-        vertSizer.Add(horzSizerT   ,1, flag = wx.LEFT|wx.EXPAND,border = 1)
-        vertSizer.Add(horzSizer    ,1, flag = wx.LEFT|wx.EXPAND,border = 1)
-
-        self.sizer = wx.BoxSizer(wx.HORIZONTAL)
-        self.sizer.Add(btSizer      ,0, flag = wx.LEFT          ,border = 1)
-        self.sizer.Add(vertSizer    ,1, flag = wx.EXPAND|wx.LEFT          ,border = 1)
-        self.SetSizer(self.sizer)
-
-        # --- Events
-        self.cbTabs.Bind   (wx.EVT_COMBOBOX, self.onTabChange)
-        self.cbFilters.Bind(wx.EVT_COMBOBOX, self.onSelectFilt)
-        self.Bind(wx.EVT_SPINCTRLDOUBLE, self.onParamChangeArrow, self.tParam)
-        self.Bind(wx.EVT_TEXT_ENTER,     self.onParamChangeEnter, self.tParam)
-        try:
-            self.spintxt = self.tParam.Children[0]
-        except:
-            self.spintxt = None
-        if platform.system()=='Windows':
-            # See issue https://github.com/wxWidgets/Phoenix/issues/1762
-            assert isinstance(self.spintxt, wx.TextCtrl)
-            self.spintxt.Bind(wx.EVT_CHAR_HOOK, self.onParamChangeChar)
-
-        # --- Init triggers
-        self.updateTabList()
-        self.onSelectFilt()
-        self.onToggleApply(init=True)
-
-    def destroy(self,event=None):
-        super(FilterToolPanel,self).destroy()
-
-    def onSelectFilt(self, event=None):
-        """ Select the filter, but does not applied it to the plotData 
-            parentFilt is unchanged
-            But if the parent already has 
-        """
-        iFilt = self.cbFilters.GetSelection()
-        filt = self._DEFAULT_FILTERS[iFilt]
-        self.lbParamName.SetLabel(filt['paramName']+':')
-        #self.tParam.SetRange(filt['paramRange'][0], filt['paramRange'][1])
-        # NOTE: if min value for range is not 0, the Ctrl prevents you to enter 0.01
-        self.tParam.SetRange(0, filt['paramRange'][1])
-        self.tParam.SetIncrement(filt['increment'])
-        self.tParam.SetDigits(filt['digits'])
-
-        parentFilt=self.parent.plotDataOptions['Filter']
-        # Value
-        if type(parentFilt)==dict and parentFilt['name']==filt['name']:
-            self.tParam.SetValue(parentFilt['param'])
-        else:
-            self.tParam.SetValue(filt['param'])
-        # Trigger plot if applied
-        self.onParamChange(self)
-
-    def _GUI2Data(self):
-        iFilt = self.cbFilters.GetSelection()
-        filt = self._DEFAULT_FILTERS[iFilt].copy()
-        try:
-            filt['param']=np.float(self.spintxt.Value)
-        except:
-            print('[WARN] pyDatView: Issue on Mac: GUITools.py/_GUI2Data. Help needed.')
-            filt['param']=np.float(self.tParam.Value)
-        if filt['param']<filt['paramRange'][0]:
-            filt['param']=filt['paramRange'][0]
-            self.tParam.SetValue(filt['paramRange'][0])
-
-        return filt
-
-    def onToggleApply(self, event=None, init=False):
-        """
-        apply Filter based on GUI Data
-        """
-        parentFilt=self.parent.plotDataOptions['Filter']
-        if not init:
-            self._filterApplied = not self._filterApplied
-
-        if self._filterApplied:
-            self.parent.plotDataOptions['Filter'] =self._GUI2Data()
-            #print('Apply', self.parent.plotDataOptions['Filter'])
-            self.lbInfo.SetLabel(
-                    'Filter is now applied on the fly. Change parameter live. Click "Clear" to stop. '
-                    )
-            self.btPlot.Enable(False)
-            self.btClear.Enable(False)
-            self.btApply.SetLabel(CHAR['sun']+' Clear')
-        else:
-            self.parent.plotDataOptions['Filter'] = None
-            self.lbInfo.SetLabel(
-                    'Click on "Apply" to set filter on the fly for all plots. '+
-                    'Click on "Plot" to try a filter on the current plot.'
-                    )
-            self.btPlot.Enable(True)
-            self.btClear.Enable(True)
-            self.btApply.SetLabel(CHAR['cloud']+' Apply')
-
-        if not init:
-            self.parent.load_and_draw() # Data will change
-            self.updateTabList()
-
-    def onAdd(self,event=None):
-        iSel         = self.cbTabs.GetSelection()
-        tabList      = self.parent.selPanel.tabList
-        icol, colname = self.parent.selPanel.xCol
-        opt = self._GUI2Data()
-        errors=[]
-        if iSel==0:
-            dfs, names, errors = tabList.applyFiltering(icol, opt, bAdd=True)
-            self.parent.addTables(dfs,names,bAdd=True)
-        else:
-            df, name = tabList.get(iSel-1).applyFiltering(icol, opt, bAdd=True)
-            self.parent.addTables([df], [name], bAdd=True)
-        self.updateTabList()
-
-        if len(errors)>0:
-            raise Exception('Error: The resampling failed on some tables:\n\n'+'\n'.join(errors))
-
-        # We stop applying
-        self.onToggleApply()
-
-    def onPlot(self, event=None):
-        """ 
-        Overlay on current axis the filter
-        """
-        from pydatview.tools.signal_analysis import applyFilter
-        if len(self.parent.plotData)!=1:
-            Error(self,'Plotting only works for a single plot. Plot less data.')
-            return
-        filt=self._GUI2Data()
-
-        PD = self.parent.plotData[0]
-        y_filt = applyFilter(PD.x0, PD.y0, filt)
-        ax = self.parent.fig.axes[0]
-
-        PD_new = PlotData()
-        PD_new.fromXY(PD.x0, y_filt)
-        self.parent.transformPlotData(PD_new)
-        ax.plot(PD_new.x, PD_new.y, '-')
-        self.parent.canvas.draw()
-
-    def onClear(self, event):
-        self.parent.load_and_draw() # Data will change
-
-    def onParamChange(self, event=None):
-        if self._filterApplied:
-            self.parent.plotDataOptions['Filter'] =self._GUI2Data()
-            #print('OnParamChange', self.parent.plotDataOptions['Filter'])
-            self.parent.load_and_draw() # Data will change
-
-    def onParamChangeArrow(self, event):
-        self.onParamChange()
-        event.Skip()
-
-    def onParamChangeEnter(self, event):
-        self.onParamChange()
-        event.Skip()
-
-    def onParamChangeChar(self, event):
-        event.Skip()  
-        code = event.GetKeyCode()
-        if code in [wx.WXK_RETURN, wx.WXK_NUMPAD_ENTER]:
-            #print(self.spintxt.Value)
-            self.tParam.SetValue(self.spintxt.Value)
-            self.onParamChangeEnter(event)
-
-    def onTabChange(self,event=None):
-        #tabList = self.parent.selPanel.tabList
-        #iSel=self.cbTabs.GetSelection()
-        pass
-
-    def updateTabList(self,event=None):
-        tabList = self.parent.selPanel.tabList
-        tabListNames = ['All opened tables']+tabList.getDisplayTabNames()
-        try:
-            iSel=np.max([np.min([self.cbTabs.GetSelection(),len(tabListNames)]),0])
-            self.cbTabs.Clear()
-            [self.cbTabs.Append(tn) for tn in tabListNames]
-            self.cbTabs.SetSelection(iSel)
-        except RuntimeError:
-            pass
-
-    def onHelp(self,event=None):
-        Info(self,"""Filtering.
-
-The filtering operation changes the "y" values of a table/plot, 
-applying a given filter (typically cutting off some frequencies).
-
-To filter perform the following step:
-
-- Chose a filtering method:
-   - Moving average: apply a moving average filter, with
-          a length specified by the window size (in indices)
-   - High pass 1st order: apply a first oder high-pass filter,
-          passing the frequencies above the cutoff frequency parameter.
-   - Low pass 1st order: apply a first oder low-pass filter,
-          passing the frequencies below the cutoff frequency parameter.
-
-- Click on one of the following buttons:
-   - Plot: will display the filtered data on the figure
-   - Apply: will perform the filtering on the fly for all new plots
-   - Add: will create new table(s) with filtered values for all 
-          signals. This process might take some time.
-          Currently done for all tables.
-""")
-
-
-
-
-# --------------------------------------------------------------------------------}
 # --- Mask
 # --------------------------------------------------------------------------------{
 class MaskToolPanel(GUIToolPanel):
@@ -937,7 +669,6 @@ Buttons:
 TOOLS={
  'LogDec':    LogDecToolPanel,
  'Outlier':   OutlierToolPanel,
- 'Filter':    FilterToolPanel,
  'Mask':      MaskToolPanel, 
  'FASTRadialAverage': RadialToolPanel,
  'CurveFitting': CurveFitToolPanel,
