+++ pydatview/plugins/__init__.py
@@ -26,12 +26,17 @@ def _data_binning(label, mainframe):
     from .data_binning import binningAction
     return binningAction(label, mainframe)
 
+def _data_filter(label, mainframe):
+    from .data_filter import filterAction
+    return filterAction(label, mainframe)
+
 def _data_sampler(label, mainframe):
     from .data_sampler import samplerAction
     return samplerAction(label, mainframe)
 
 dataPlugins=[
         # Name/label             , callback                , is a Panel
+        ('Filter'                , _data_filter            , True ),
         ('Resample'              , _data_sampler           , True ),
         ('Bin data'              , _data_binning           , True ),
         ('Standardize Units (SI)', _data_standardizeUnitsSI, False),
